# ManglerAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ManglerAPI v1",
    "description": "REST API for StoryMangler",
    "version": "v1"
  },
  "paths": {
    "/helloworld": {
      "get": {
        "tags": [
          "HelloWorld"
        ],
        "operationId": "GetHelloWorld",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HelloWorldModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelloWorldModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelloWorldModel"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/exceptionTest": {
      "get": {
        "tags": [
          "HelloWorld"
        ],
        "operationId": "GetHelloWorld",
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/story/{id}": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Retrieve a single story by ID.  Will not include story lines.",
        "operationId": "GetStory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/stories": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Returns a list of stories matching the specified criteria.  Stories are returned from newest to oldest.",
        "operationId": "GetStory",
        "parameters": [
          {
            "name": "GuildIdentifier",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AuthorIdentifier",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cursor",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModelCursorPaginatedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModelCursorPaginatedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModelCursorPaginatedList"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Story"
        ],
        "summary": "Creates a new story.",
        "operationId": "PostStory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryPostRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryPostRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StoryPostRequestModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryModel"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClientTypeModel": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Error code representing the specific error.",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Detailed message of the error.",
            "nullable": true
          },
          "callId": {
            "type": "string",
            "description": "The callId of the errored response.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HelloWorldModel": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StoryModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The primary key of the Story record",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "The title of the story",
            "nullable": true
          },
          "authorUserIdentifier": {
            "type": "string",
            "description": "UserIdentifier of the story's author.",
            "nullable": true
          },
          "authorClientType": {
            "$ref": "#/components/schemas/ClientTypeModel"
          },
          "guildIdentifier": {
            "type": "string",
            "description": "Group that is responsible for this story, or null if it is open to the world.",
            "nullable": true
          },
          "completeDt": {
            "type": "string",
            "description": "Date the story was marked completed, or null if the story is still open.",
            "format": "date-time",
            "nullable": true
          },
          "createDt": {
            "type": "string",
            "description": "Date that the story was first created.",
            "format": "date-time"
          },
          "isViewableByPublic": {
            "type": "boolean",
            "description": "If true, the story will show up to the public"
          },
          "isWritableByPublic": {
            "type": "boolean",
            "description": "If true, will show up to users who want to contribute to a random story"
          }
        },
        "additionalProperties": false
      },
      "StoryModelCursorPaginatedList": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoryModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StoryPostRequestModel": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the story",
            "nullable": true
          },
          "authorUserIdentifier": {
            "type": "string",
            "description": "UserIdentifier of the story's author.",
            "nullable": true
          },
          "authorClientType": {
            "$ref": "#/components/schemas/ClientTypeModel"
          },
          "guildIdentifier": {
            "type": "string",
            "description": "Group that is responsible for this story, or null if it is open to the world.",
            "nullable": true
          },
          "createDt": {
            "type": "string",
            "description": "Date that the story was first created.",
            "format": "date-time"
          },
          "isViewableByPublic": {
            "type": "boolean",
            "description": "If true, the story will show up to the public"
          },
          "isWritableByPublic": {
            "type": "boolean",
            "description": "If true, will show up to users who want to contribute to a random story"
          },
          "firstLine": {
            "type": "string",
            "description": "The first line of the story to kick things off.",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}